ask-multiple-pdfs/requirements.txt
langchain== 1.0.2
langchain-community== 0.4.1
langchain-google-genai== 3.0.0
chromadb==1.3.0

# google-generativeai==0.7.2

python-dotenv
streamlit
pypdf
fastapi
uvicorn

tiktoken
sentence-transformers
